requirements.txt
flask
flask-cors
pymysql
python-dotenv
werkzeug
request